CRUD-project updates

+++ crud/students/views.py
@@ -1,5 +1,5 @@
 from django.shortcuts import render
-from django.http import HttpRequest, HttpResponse, JsonResponse
+from django.http import HttpRequest, HttpResponse, JsonResponse, Http404
 from django.views.decorators.csrf import csrf_exempt
 
 from .models import Student
@@ -25,17 +25,33 @@ def all_students(request):
                   context={'students': all_students_list})
 
 
-@csrf_exempt
 def add_student(request):
-    # print(request.body)
-    student_data = loads(request.body.decode('utf-8'))
-    print(student_data)
-    new_student = Student(name=student_data['name'], mark=student_data['mark'])
-    new_student.save()
-    return JsonResponse({'ok': True, 'data': student_data})
+    if request.method == 'POST':
+        student_data = loads(request.body.decode('utf-8'))
+        new_student = Student(name=student_data['name'], mark=student_data['mark'])
+        new_student.save()
+        return JsonResponse({'ok': True, 'data': student_data})
+    else:
+        raise Http404()
 
 
-@csrf_exempt
 def delete_student(request, student_id):
-    Student.objects.get(id=student_id).delete()
-    return JsonResponse({'ok': True})
+    if request.method == 'POST':
+        Student.objects.get(id=student_id).delete()
+        return JsonResponse({'ok': True})
+    else:
+        raise Http404()
+
+
+def update_student(request):
+    if request.method == 'POST':
+        student_data = loads(request.body.decode('utf-8'))
+        name = student_data['name']
+        mark = student_data['mark']
+        student = Student.objects.get(id=student_data['id'])
+        student.name = name
+        student.mark = mark
+        student.save()
+        return JsonResponse({'ok': True})
+    else:
+        raise Http404()
